+++ apps/microsoft-teams/frontend/src/components/config/AddButton/AddButton.spec.tsx
@@ -4,8 +4,17 @@ import { render, screen } from '@testing-library/react';
 
 describe('AddButton component', () => {
   it('mounts with correct button copy', () => {
-    render(<AddButton buttonCopy="test" handleClick={vi.fn()} />);
+    const { unmount } = render(<AddButton buttonCopy="test" handleClick={vi.fn()} />);
 
     expect(screen.getByText('test')).toBeTruthy();
+    unmount();
+  });
+  it('is disabled when prop is true', () => {
+    const { unmount } = render(
+      <AddButton buttonCopy="test" handleClick={vi.fn()} isDisabled={true} />
+    );
+
+    expect(screen.getByRole('button')).toHaveProperty('disabled');
+    unmount();
   });
 });