+++ apps/microsoft-teams/frontend/src/components/config/AddButton/AddButton.tsx
@@ -5,13 +5,18 @@ import { PlusIcon } from '@contentful/f36-icons';
 interface Props {
   buttonCopy: string;
   handleClick: (e: MouseEvent<HTMLButtonElement>) => void;
+  isDisabled?: boolean;
 }
 
 const AddButton = (props: Props) => {
-  const { buttonCopy, handleClick } = props;
+  const { buttonCopy, handleClick, isDisabled = false } = props;
 
   return (
-    <Button startIcon={<PlusIcon />} variant="secondary" onClick={handleClick}>
+    <Button
+      startIcon={<PlusIcon />}
+      variant="secondary"
+      onClick={handleClick}
+      isDisabled={isDisabled}>
       {buttonCopy}
     </Button>
   );