+++ apps/microsoft-teams/frontend/src/components/config/ContentTypeSelection/ContentTypeSelection.tsx
@@ -1,3 +1,4 @@
+import { useContext } from 'react';
 import { Box, Flex, IconButton, ModalLauncher, Text, TextInput } from '@contentful/f36-components';
 import AddButton from '@components/config/AddButton/AddButton';
 import ContentTypeSelectionModal from '@components/config/ContentTypeSelectionModal/ContentTypeSelectionModal';
@@ -6,17 +7,17 @@ import { contentTypeSelection } from '@constants/configCopy';
 import { Notification } from '@customTypes/configPage';
 import { EditIcon } from '@contentful/f36-icons';
 import { styles } from './ContentTypeSelection.styles';
-import { ContentTypeProps } from 'contentful-management';
 import { getContentTypeName } from '@helpers/configHelpers';
+import { ContentTypeContext } from '@context/ContentTypeProvider';
 
 interface Props {
   notification: Notification;
   handleNotificationEdit: (notificationEdit: Partial<Notification>) => void;
-  contentTypes: ContentTypeProps[];
 }
 
 const ContentTypeSelection = (props: Props) => {
-  const { notification, handleNotificationEdit, contentTypes } = props;
+  const { notification, handleNotificationEdit } = props;
+  const { contentTypes, contentTypeConfigLink } = useContext(ContentTypeContext);
 
   const openContentTypeSelectionModal = () => {
     return ModalLauncher.open(({ isShown, onClose }) => (
@@ -28,6 +29,7 @@ const ContentTypeSelection = (props: Props) => {
         handleNotificationEdit={handleNotificationEdit}
         savedContentTypeId={notification.contentTypeId}
         contentTypes={contentTypes}
+        contentTypeConfigLink={contentTypeConfigLink}
       />
     ));
   };