+++ apps/microsoft-teams/frontend/src/components/config/ContentTypeSelectionModal/ContentTypeSelectionModal.spec.tsx
@@ -12,6 +12,7 @@ describe('ContentTypeSelectionModal component', () => {
         savedContentTypeId=""
         handleNotificationEdit={vi.fn()}
         contentTypes={[]}
+        contentTypeConfigLink=""
       />
     );