+++ apps/microsoft-teams/frontend/src/components/config/DeleteModal/DeleteModal.spec.tsx
@@ -1,12 +1,12 @@
 import DeleteModal from './DeleteModal';
 import { describe, expect, it, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { editModeFooter } from '@constants/configCopy';
+import { notificationsSection } from '@constants/configCopy';
 
 describe('DeleteModal component', () => {
   it('mounts and renders the correct content', () => {
     render(<DeleteModal isShown={true} handleCancel={vi.fn()} handleDelete={vi.fn()} />);
 
-    expect(screen.getByText(editModeFooter.confirmDelete)).toBeTruthy();
+    expect(screen.getByText(notificationsSection.confirmDelete)).toBeTruthy();
   });
 });