+++ apps/microsoft-teams/frontend/src/components/config/NotificationViewMode/NotificationViewMode.tsx
@@ -1,8 +1,18 @@
-import { Box, Button, Flex, Subheading, Paragraph, Switch } from '@contentful/f36-components';
+import { useContext } from 'react';
+import { ContentTypeContext } from '@context/ContentTypeProvider';
+import {
+  Box,
+  IconButton,
+  Flex,
+  Menu,
+  Subheading,
+  Paragraph,
+  Switch,
+} from '@contentful/f36-components';
+import { MoreHorizontalIcon } from '@contentful/f36-icons';
 import { styles } from './NotificationViewMode.styles';
 import { getContentTypeName, getChannelName } from '@helpers/configHelpers';
 import { Notification } from '@customTypes/configPage';
-import { ContentTypeProps } from 'contentful-management';
 import {
   channelSelection,
   contentTypeSelection,
@@ -15,14 +25,15 @@ interface Props {
   index: number;
   updateNotification: (index: number, editedNotification: Partial<Notification>) => void;
   notification: Notification;
-  contentTypes: ContentTypeProps[];
   handleEdit: () => void;
-  isEditDisabled: boolean;
+  isMenuDisabled: boolean;
+  handleDelete: () => void;
 }
 
 const NotificationViewMode = (props: Props) => {
-  const { index, notification, updateNotification, contentTypes, handleEdit, isEditDisabled } =
+  const { index, notification, updateNotification, handleEdit, isMenuDisabled, handleDelete } =
     props;
+  const { contentTypes } = useContext(ContentTypeContext);
 
   return (
     <Box className={styles.wrapper}>
@@ -50,13 +61,21 @@ const NotificationViewMode = (props: Props) => {
             onChange={() => updateNotification(index, { isEnabled: !notification.isEnabled })}
           />
           <Box marginLeft="spacingXs">
-            <Button
-              variant="secondary"
-              size="small"
-              onClick={handleEdit}
-              isDisabled={isEditDisabled}>
-              {notificationsSection.editButton}
-            </Button>
+            <Menu>
+              <Menu.Trigger>
+                <IconButton
+                  testId="menu-button"
+                  variant="transparent"
+                  icon={<MoreHorizontalIcon />}
+                  aria-label="toggle menu"
+                  isDisabled={isMenuDisabled}
+                />
+              </Menu.Trigger>
+              <Menu.List>
+                <Menu.Item onClick={handleEdit}>{notificationsSection.edit}</Menu.Item>
+                <Menu.Item onClick={handleDelete}>{notificationsSection.delete}</Menu.Item>
+              </Menu.List>
+            </Menu>
           </Box>
         </Flex>
       </Flex>