+++ apps/microsoft-teams/frontend/src/components/config/NotificationsSection/NotificationsSection.tsx
@@ -1,14 +1,15 @@
 import { Dispatch, useState } from 'react';
-import { Box, Subheading } from '@contentful/f36-components';
+import { Box, ModalLauncher, Subheading } from '@contentful/f36-components';
 import { styles } from './NotificationsSection.styles';
 import { notificationsSection } from '@constants/configCopy';
 import AddButton from '@components/config/AddButton/AddButton';
 import NotificationEditMode from '@components/config/NotificationEditMode/NotificationEditMode';
 import NotificationViewMode from '@components/config/NotificationViewMode/NotificationViewMode';
+import DeleteModal from '@components/config/DeleteModal/DeleteModal';
 import { Notification } from '@customTypes/configPage';
 import { ParameterAction, actions } from '@components/config/parameterReducer';
-import useGetContentTypes from '@hooks/useGetContentTypes';
 import useGetTeamsChannels from '@hooks/useGetTeamsChannels';
+import { ContentTypeContextProvider } from '@context/ContentTypeProvider';
 
 interface Props {
   notifications: Notification[];
@@ -18,7 +19,6 @@ interface Props {
 const NotificationsSection = (props: Props) => {
   const { notifications, dispatch } = props;
   const [notificationIndexToEdit, setNotificationIndexToEdit] = useState<number | null>(null);
-  const contentTypes = useGetContentTypes();
   const channels = useGetTeamsChannels();
 
   const createNewNotification = () => {
@@ -32,6 +32,24 @@ const NotificationsSection = (props: Props) => {
     dispatch({
       type: actions.UPDATE_NOTIFICATIONS,
       payload: notificationsPayload,
+    });
+  };
+
+  const handleDelete = (index: number) => {
+    ModalLauncher.open(({ isShown, onClose }) => {
+      return (
+        <DeleteModal
+          isShown={isShown}
+          handleCancel={() => {
+            onClose(true);
+          }}
+          handleDelete={() => {
+            onClose(true);
+            deleteNotification(index);
+            setNotificationIndexToEdit(null);
+          }}
+        />
+      );
     });
   };
 
@@ -51,38 +69,40 @@ const NotificationsSection = (props: Props) => {
         <AddButton
           buttonCopy={notificationsSection.createButton}
           handleClick={createNewNotification}
+          isDisabled={notificationIndexToEdit !== null}
         />
       </Box>
-      {notifications.map((notification, index) => {
-        const inEditMode = notificationIndexToEdit === index;
+      <ContentTypeContextProvider>
+        {notifications.map((notification, index) => {
+          const inEditMode = notificationIndexToEdit === index;
 
-        if (inEditMode) {
-          return (
-            <NotificationEditMode
-              key={`notification-${index}`}
-              index={index}
-              deleteNotification={deleteNotification}
-              updateNotification={updateNotification}
-              notification={notification}
-              contentTypes={contentTypes}
-              setNotificationIndexToEdit={setNotificationIndexToEdit}
-              channels={channels}
-            />
-          );
-        } else {
-          return (
-            <NotificationViewMode
-              key={`notification-${index}`}
-              index={index}
-              updateNotification={updateNotification}
-              notification={notification}
-              contentTypes={contentTypes}
-              handleEdit={() => setNotificationIndexToEdit(index)}
-              isEditDisabled={notificationIndexToEdit !== null}
-            />
-          );
-        }
-      })}
+          if (inEditMode) {
+            return (
+              <NotificationEditMode
+                key={`notification-${index}`}
+                index={index}
+                deleteNotification={deleteNotification}
+                updateNotification={updateNotification}
+                notification={notification}
+                setNotificationIndexToEdit={setNotificationIndexToEdit}
+                channels={channels}
+              />
+            );
+          } else {
+            return (
+              <NotificationViewMode
+                key={`notification-${index}`}
+                index={index}
+                updateNotification={updateNotification}
+                notification={notification}
+                handleEdit={() => setNotificationIndexToEdit(index)}
+                isMenuDisabled={notificationIndexToEdit !== null}
+                handleDelete={() => handleDelete(index)}
+              />
+            );
+          }
+        })}
+      </ContentTypeContextProvider>
     </Box>
   );
 };