+++ packages/ecommerce-app-base/src/ProductCard/ProductCard/ProductCard.tsx
@@ -71,7 +71,7 @@ export const ProductCard = (props: ProductCardProps) => {
         <TextLink
           as="button"
           onClick={() => setIsExpanded((currentIsExpanded) => !currentIsExpanded)}>
-          {isExpanded ? 'hide details' : 'more details'}
+          {isExpanded ? 'Hide details' : 'More details'}
         </TextLink>
       </Box>
     );